Fixes miss shader
This commit fixes the miss shader issue, I swapped the order in which we bind the shaders in the shader binding table and fixed the python script to generate the shader with the correct arguments. Now when setting up the raytracing pipeline, we pass the correct shaders. Finally i cleaned up the engine by removing unused file and code.