Renders the mesh's grey color and miss/hit fix
The Hit and Miss shaders somehow were flipped again, made them consistent around the engine now so it is always correct. Also found a way to retrieve the material uniform buffer which holds the sRGB and passed this to the shader, correctly rendering the base grey color of the mesh.