Albedo Rendering
The pipeline now correctly retrieves the materials and stores them in a LocalVector, which is indexed in the shader to retrieve the correct material data for each mesh. Also fixes the issue where there is a helper plane, geometry piece in the origin of the scene. This plane has no mesh which broke some things, and got a BLAS created for it, which broke material indexing.